+++ packages/typespec-rust/test/cadlranch/payload/content-negotiation/tests/content_negotiation_different_body_test.rs
@@ -30,7 +30,7 @@ async fn get_avatar_as_png() {
         .get_avatar_as_png(None)
         .await
         .unwrap();
-    let body = resp.into_body().collect().await.unwrap();
+    let body = resp.into_raw_body().collect().await.unwrap();
     let image_png =
         fs::read("../../../../node_modules/@azure-tools/cadl-ranch-specs/assets/image.png")
             .unwrap();